Simple Code for Quick select for k th smallest Element
It repeatedly calls the pivot swap function to partition the array until the kth smallest element is found. It adjusts the search range based on the pivot's position relative to k. If the pivot is at index k - 1, it returns the element at that index. If the pivot is to the left of k - 1, it searches in the left subarray. If the pivot is to the right of k - 1, it searches in the right subarray. (Simple Code without complexity for any random n elements which will be input by user itself)