Iterative Quick Sort
partition Function: This function takes the last element of the array as the pivot, places the pivot element at its correct position in the sorted array, and places all smaller elements (smaller than the pivot) to the left of the pivot and all greater elements to the right. It returns the index of the pivot element.

iterativeQuickSort Function: This function performs Quick Sort iteratively using a stack. It maintains a stack to store the indices of subarrays that need to be sorted. It iteratively partitions the array until all subarrays are sorted. It counts the number of partition calls made during the sorting process.