Activity Selection
Activity Comparison Function:  activityCompare is a comparison function used to sort activities based on their finish times. It returns true if the finish time of activity a is less than the finish time of activity 

Random Activity Generation:   The generateRandomActivities function fills an array of Activity structures with random start and finish times. The start time is randomly generated, and the finish time is set to be after the start time.