Prim's Algorithm for Minimum Spanning Tree
code 61 is random example of a 4 - 4 matrix graph and code is 62 is based on 2D array graph[][] , considering the number of edges m and the number of vertices n as input. User can input any sort of desired values and check for the results and thus gives the minimum spanning tree for the graph.